+++ src/cli/middleware/set-logger.js
@@ -1,0 +1,5 @@
+const Logger = require('../../utils/logger')
+
+module.exports = (argv) => {
+  argv.logger = new Logger(argv)
+}