+++ src/utils/create-new-project/add-deployment-target/index.js
@@ -1,0 +1,12 @@
+const targets = require('./targets')
+const addDocker = require('./docker')
+
+module.exports = async (cwd, options) => {
+  const target = options.deploymentTarget
+
+  if (!targets.includes(target)) {
+    throw new Error(`Deployment target not supported: ${target}`)
+  } else if (target === 'docker') {
+    return addDocker(cwd)
+  }
+}