+++ src/utils/create-new-project/index.js
@@ -2,11 +2,11 @@ const path = require('path')
 const fs = require('fs-extra')
 const updateProjectConfig = require('../update-project-config')
 const setupGit = require('./setup-git')
-const log = require('../log')
 const { installDependencies } = require('../manage-dependencies')
 const addComponents = require('./add-components')
 const updatePackage = require('./update-package')
 const updateKoopConfig = require('./update-koop-config')
+const addDeploymentTarget = require('./add-deployment-target')
 
 /**
  * Creat a new koop project.
@@ -22,6 +22,7 @@ const updateKoopConfig = require('./update-koop-config')
  */
 module.exports = async (cwd, type, name, options = {}) => {
   const projectPath = path.join(cwd, name)
+  const { logger } = options
 
   /**
    * Create project directory and copy project template
@@ -39,7 +40,7 @@ module.exports = async (cwd, type, name, options = {}) => {
   await updatePackage(projectPath, type, name)
   await updateKoopConfig(projectPath, type, name, options)
 
-  log('\u2713 created project', 'info', options)
+  logger.info('\u2713 created project')
 
   /**
    * Initialize Git
@@ -47,7 +48,7 @@ module.exports = async (cwd, type, name, options = {}) => {
 
   if (!options.skipGit) {
     await setupGit(projectPath)
-    log('\u2713 initialized Git', 'info', options)
+    logger.info('\u2713 initialized Git')
   }
 
   /**
@@ -55,7 +56,7 @@ module.exports = async (cwd, type, name, options = {}) => {
    */
 
   await updateProjectConfig(projectPath, type, name, options.config)
-  log(`\u2713 added ${type} configuration`, 'info', options)
+  logger.info(`\u2713 added ${type} configuration`)
 
   /**
    * Install dependencies
@@ -64,12 +65,17 @@ module.exports = async (cwd, type, name, options = {}) => {
   if (!options.skipInstall) {
     // install dependencies
     await installDependencies(projectPath, options)
-    log('\u2713 installed dependencies', 'info', options)
+    logger.info('\u2713 installed dependencies')
+  }
+
+  if (options.addDeploymentTarget) {
+    // add deployment target addon files
+    await addDeploymentTarget(projectPath, options)
   }
 
   /**
    * Done
    */
 
-  log('\u2713 done', 'info', options)
+  logger.info('\u2713 done')
 }