+++ src/utils/update-project-config.js
@@ -1,9 +1,17 @@
 const path = require('path')
 const fs = require('fs-extra')
+const _ = require('lodash')
 const writeJson = require('./write-formatted-json')
 
 module.exports = async (cwd, type, name, newConfig = {}) => {
   const configPath = path.join(cwd, 'config/default.json')
+
+  // do nothing if
+  // 1. the app is not using config/defult.json for configuration
+  // 2. the new config is empty
+  if (!(await fs.pathExists(configPath)) || _.isEmpty(newConfig)) {
+    return
+  }
 
   let config = {}